added missing files to git
done more work on unit testing. Current unit test on checks returned values and that they match the returned values from swf. Need to then submit results to google spreadsheet so I can keep a record of the results.

Need to create a second unit test, that checks the visuals match - some kind of slide div on top of swf - not sure. :D